+++ src/postprocessing_func.h
@@ -1,0 +1,13 @@
+#ifndef LBM_POSTPROCESSING_FUNC_H
+#define LBM_POSTPROCESSING_FUNC_H
+
+namespace pp{
+enum class FuncType{LINE};
+};
+
+template<GInt NDIM>
+class PostprocessFunctionExecutorInterface{
+
+};
+
+#endif // LBM_POSTPROCESSING_FUNC_H